[Keymap] Jonavin userspace encoder alt tab scrolling (#14979)
