+++ keyboards/mechwild/obe/keymaps/jonavin/rules.mk
@@ -10,3 +10,5 @@ STARTUP_NUMLOCK_ON = yes
 ENCODER_DEFAULTACTIONS_ENABLE = yes
 EMOTICON_ENABLE = yes
 INVERT_NUMLOCK_INDICATOR = yes
+
+ALTTAB_SCROLL_ENABLE = yes
