+++ users/jonavin/jonavin.h
@@ -45,6 +45,8 @@ enum custom_user_keycodes {
   EMO_JOY,      // (^o^)
   EMO_TEARS,    // (T_T)
 
+  KC_TSTOG,     // Tab Scroll Toggle
+
   NEW_SAFE_RANGE  // new safe range for keymap level custom keycodes
 };
 
@@ -91,6 +93,12 @@ enum custom_user_keycodes {
         void encoder_action_rgb_brightness(bool clockwise);
         void encoder_action_rgb_mode(bool clockwise);
     #endif // RGB_MATRIX_ENABLE / RGBLIGHT_ENABLE
+
+    #ifdef ALTTAB_SCROLL_ENABLE
+        void encoder_action_alttabscroll(bool clockwise);
+        void encoder_toggle_alttabscroll(void);
+        void encoder_tick_alttabscroll(void);
+    #endif // ALTTAB_SCROLL_ENABLE
 #endif // ENCODER_ENABLE
