+++ users/jonavin/rules.mk
@@ -1,6 +1,13 @@
 SRC += jonavin.c
-ifeq ($(strip $(ENCODER_DEFAULTACTIONS_ENABLE)), yes)
-    OPT_DEFS += -DENCODER_DEFAULTACTIONS_ENABLE
+ifdef ENCODER_ENABLE
+	# include encoder related code when enabled
+	ifeq ($(strip $(ENCODER_DEFAULTACTIONS_ENABLE)), yes)
+		OPT_DEFS += -DENCODER_DEFAULTACTIONS_ENABLE
+	endif
+	ifeq ($(strip $(ALTTAB_SCROLL_ENABLE)), yes)
+		OPT_DEFS += -DALTTAB_SCROLL_ENABLE
+	endif
+	SRC += jonavin_encoder.c
 endif
 ifeq ($(strip $(TD_LSFT_CAPSLOCK_ENABLE)), yes)
     OPT_DEFS += -DTD_LSFT_CAPSLOCK_ENABLE
